samples/github-helper-agent/pyproject.toml
[project]
name = "github-helper-agent"
version = "0.0.8"
description = "An automated agent that reviews GitHub pull requests and provides feedback"
authors = [{ name = "Cristi Pufu", email = "cristian.pufu@uipath.com" }]
dependencies = [
    "uipath-langchain==0.0.93",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.0.1",
    "langgraph>=0.3.21",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "anthropic>=0.8.0",
    "langchain-mcp-adapters>=0.0.9"
]
requires-python = ">=3.10"
